refactor: move actions into proto::v1
