+++ src/modules/mod.rs
@@ -1,2 +1,1 @@
 pub mod chassis;
-pub mod sound;
